Added a game over state and a background.
The game over state for this lesson is very simplistic, it is literally just a function that resets the scene when the player falls below a specific global position on the y axis.

For the background I have again opted to not follow the lesson, because I don't want to change the background for my editor. So instead I created a second tile map and tile set using a the background tiles from https://kenney.nl/assets/pixel-platformer. I think the results are very good looking compared to a simple colour change.