Added an enemy character.
Following the lesson I have added a basic enemy that glides between two positions.

The major difference from the lesson is the way I have the enemy switch which location is the one that they are to go after reaching the previous one. In the lesson they have used if statements to determine which point is to be used for the new target position. I deemed that a bit inefficient so I simple used a temporary variable to switch the start and end positions to minimally reduce the CPU load at the expense of temporary RAM use.