Completed course.
I have spent a bit too much time on another unnecessary addition to the project. I have added a terrain definition to the tile set to make it easier to draw new levels.

Otherwise an end flag was added so that the player can progress to the next level. Also level 2 was created.